[REF] pylint.conf: Consider our partner as possible required author
The lint to detect required authors in modules's manifest
(manifest-required-author) now accepts multiple possible values, so this
commit includes our partner on those.

This is similar to e1457a0bf556, which had to be reverted, because it
was applied before the lint improvement

Closes https://github.com/Vauxoo/pylint-odoo/issues/157

Note:
This change was already applied on 86b1ca8fa22d and reverted on
c0809a2d2604, to solve the issue Vauxoo/pylint-odoo#159. However, this
is being re-aplied because it's unrelated to the aforementioned issue,
and it's addressed on PR Vauxoo/pylint-odoo#160